backend/requirements.txt
Flask
flask-cors
Pillow
opencv-python
scikit-learn
python-multipart
requests
